Move new classes to api, fix imports
