+++ src/main/java/me/zeroeightsix/fiber/api/annotation/AnnotatedSettings.java
@@ -3,6 +3,9 @@ package me.zeroeightsix.fiber.api.annotation;
 import me.zeroeightsix.fiber.api.NodeOperations;
 import me.zeroeightsix.fiber.api.annotation.convention.SettingNamingConvention;
 import me.zeroeightsix.fiber.api.annotation.exception.MalformedFieldException;
+import me.zeroeightsix.fiber.api.exception.RuntimeFiberException;
+import me.zeroeightsix.fiber.api.tree.ConfigNode;
+import me.zeroeightsix.fiber.api.tree.ConfigTree;
 import me.zeroeightsix.fiber.impl.annotation.convention.NoNamingConvention;
 import me.zeroeightsix.fiber.api.builder.ConfigAggregateBuilder;
 import me.zeroeightsix.fiber.api.builder.ConfigTreeBuilder;
@@ -10,7 +13,6 @@ import me.zeroeightsix.fiber.api.builder.ConfigLeafBuilder;
 import me.zeroeightsix.fiber.api.builder.constraint.AbstractConstraintsBuilder;
 import me.zeroeightsix.fiber.api.exception.FiberException;
 import me.zeroeightsix.fiber.impl.annotation.magic.TypeMagic;
-import me.zeroeightsix.fiber.api.tree.ConfigBranch;
 
 import javax.annotation.Nonnull;
 import java.lang.annotation.Annotation;
